requirements.txt
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
django-mptt==0.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
psycopg2==2.9.9
environ==1.0
eskiz-sms==0.2.3
Pillow==10.1.0
djoser_passwordless